Splitting tests for better readability and debugging
