+++ tests/fixtures/MakeUserModelNoPassword/tests/GeneratedEntityTest.php
@@ -9,7 +9,7 @@ class GeneratedEntityTest extends WebTestCase
     public function testGeneratedEntity()
     {
         // login then access a protected page
-        $client = static::createClient();
+        $client = self::createClient();
         $client->request('GET', '/login?username=hal9000');
 
         $this->assertSame(302, $client->getResponse()->getStatusCode());
